backend/requirements.txt
Django==2.0.5
djangorestframework==3.8.2
django_extensions==2.0.7
psycopg2
